Renaming collect_animation to collect_curves
